# client-sdks/manager/rust/openapi-3.0.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Alien Manager API",
    "description": "Control plane for Alien applications. Manages deployments, releases, commands, and telemetry.",
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/initialize": {
      "post": {
        "tags": [
          "sync"
        ],
        "operationId": "initialize",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitializeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Existing deployment returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InitializeResponse"
                }
              }
            }
          },
          "201": {
            "description": "New deployment created with token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InitializeResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "InitializeRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "platform": {
            "$ref": "#/components/schemas/Platform",
            "nullable": true
          }
        }
      },
      "InitializeResponse": {
        "type": "object",
        "required": [
          "deploymentId"
        ],
        "properties": {
          "deploymentId": {
            "type": "string"
          },
          "token": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "Platform": {
        "type": "string",
        "description": "Represents the target cloud platform.",
        "enum": [
          "aws",
          "gcp",
          "azure",
          "kubernetes",
          "local",
          "test"
        ]
      },
...
    }
  },
...
}